requirements.txt
asgiref==3.8.1
Django==4.2
django-environ==0.12.0
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
